+++ libdariadb/interfaces/imeassource.h
@@ -6,6 +6,7 @@
 #include <libdariadb/st_exports.h>
 #include <libdariadb/status.h>
 #include <libdariadb/query_param.h>
+#include <libdariadb/stat.h>
 #include <memory>
 
 namespace dariadb {
@@ -19,10 +20,11 @@ public:
   virtual void foreach (const QueryInterval &q, IReadCallback * clbk) = 0;
   virtual void foreach (const QueryTimePoint &q, IReadCallback * clbk);
   virtual Id2Cursor intervalReader(const QueryInterval &query)=0;
-  EXPORT virtual MeasList readInterval(const QueryInterval &q);
   virtual Id2Meas readTimePoint(const QueryTimePoint &q) = 0;
   virtual Id2Meas currentValue(const IdArray &ids, const Flag &flag) = 0;
+  virtual Statistic stat(const Id id, Time from, Time to)=0;
   EXPORT virtual Id2MinMax loadMinMax();
+  EXPORT virtual MeasList readInterval(const QueryInterval &q);
   virtual ~IMeasSource() {}
 };