+++ libdariadb/storage/memstorage/memstorage.h
@@ -31,6 +31,7 @@ public:
   EXPORT virtual bool minMaxTime(dariadb::Id id, dariadb::Time *minResult,
                                  dariadb::Time *maxResult) override;
   EXPORT virtual Id2Cursor intervalReader(const QueryInterval &q)override;
+  EXPORT  Statistic stat(const Id id, Time from, Time to)override;
   EXPORT virtual void foreach (const QueryInterval &q, IReadCallback * clbk) override;
   EXPORT virtual Id2Meas readTimePoint(const QueryTimePoint &q) override;
   EXPORT virtual Id2Meas currentValue(const IdArray &ids, const Flag &flag) override;